requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-google==6.2.0
google-cloud-storage==2.8.0
google-cloud-bigquery==3.5.0
pytz==2023.3
pandas==1.5.3
